Add enders.py and groomed output file from it
These are all 1024, possible, unique starters of the 10 initial (right
side) steno keys, along with their occurrence counts in main.json,
sorted from least to most occurring. Enders are counted in every stroke
of all outlines, as this gives a real count of how often each one is
used. The keys considered for every entry here are "FRPBLGTSDZ", and any
letters not in a given chord below are considered unstroked. For
example, the starter 'FRPBLT' means that the -FRPBLT keys are all
stroked, while the -GSDZ keys are not.